image-agents/src/tools-service/requirements.txt
fastapi
python-multipart
pydantic

uvicorn
sentencepiece
protobuf
librosa

torch
torchvision
pillow

pandas
opencv-python
transformers
accelerate